Add BasicClient to networking code
The BasicClient acts as a base for building an actual reversi client.
This functionality is kept seperate from game code to be able to reuse it later for different client types to collect game data.

Note: Code is still untested until BasicServer is also implemented to test them against each other.