Store board as Enum Matrix.
The numpy matrix made move execution VERY slow (factor of 4 slower).
The execution speed is not extremly important, but it should be within a reasonable magnitude, as it should not be the main factor in execution speed (this would usually be the neural network execution).

Now the board is saved as a Matrix of enum objects in a playn python 2D array.
This seems like a fair ballance between execution speed and handling of the data type for transformations.