Add basic two player evaluation against the ai_trivial client.
This allows to play a tournament between the trained neural network and the ai_trivial reversi client.
It is needed to keep an metric of the actual strength of the NN, as the real playing strength can not be evaluated when playing against itself.
The goal is to train an NN that can at least win a portion of the games with consistency.