Extend WorkerSlave to be able to do selfplay, selfevaluation and aievaluation.
This should allow very esay execution of the above tasks over many machines.
Before the selfplay took about half the time as the training (on pure cpy machines).
If a GPU is added for training it will speed up to about 10x it’s speed. With this distributed generation of sample games we can keep a good ballance between training and selfplay.